Added origen_app_generators to Origen gem package

+++ lib/origen/commands/new.rb
@@ -62,7 +62,12 @@ else
   end
 end
 
-generators = [['http://rubygems.org', 'origen_app_generators']] + Array(Origen.site_config.app_generators)
+generators = [['https://rubygems.org', 'origen_app_generators']] + Array(Origen.site_config.app_generators)
+
+def use_packaged_generators
+  FileUtils.cp_r Origen.top.join('origen_app_generators').to_s, '.'
+  FileUtils.mv 'origen_app_generators', '0'
+end
 
 if update_required
   puts 'Fetching the latest app generators...'
@@ -73,41 +78,46 @@ if update_required
     generators.each_with_index do |gen, i|
       # If a reference to a gem from a gem server
       if gen.is_a?(Array)
-        response = HTTParty.get("#{gen[0]}/api/v1/dependencies.json?gems=#{gen[1]}")
+        begin
+          response = HTTParty.get("#{gen[0]}/api/v1/dependencies.json?gems=#{gen[1]}", timeout: 2)
 
-        if response.success?
-          latest_version = JSON.parse(response.body).map { |v| v['number'] }.max
-
-          response = HTTParty.get("#{gen[0]}/gems/#{gen[1]}-#{latest_version}.gem")
           if response.success?
-            File.open("#{gen[1]}-#{latest_version}.gem", 'wb') do |f|
-              f.write response.parsed_response
+            latest_version = JSON.parse(response.body).map { |v| v['number'] }.max
+
+            response = HTTParty.get("#{gen[0]}/gems/#{gen[1]}-#{latest_version}.gem", timeout: 5)
+            if response.success?
+              File.open("#{gen[1]}-#{latest_version}.gem", 'wb') do |f|
+                f.write response.parsed_response
+              end
+
+              `gem unpack #{gen[1]}-#{latest_version}.gem`
+              FileUtils.rm_rf("#{gen[1]}-#{latest_version}.gem")
+              FileUtils.mv("#{gen[1]}-#{latest_version}", i.to_s)
+            else
+              use_packaged_generators if i == 0
             end
           else
-            puts "Sorry, could not find generator #{gen[1]} version #{latest_version}"
+            use_packaged_generators if i == 0
           end
-
-          `gem unpack #{gen[1]}-#{latest_version}.gem`
-          FileUtils.rm_rf("#{gen[1]}-#{latest_version}.gem")
-          FileUtils.mv("#{gen[1]}-#{latest_version}", i.to_s)
-
-        else
-          puts "Failed to get generator #{gen[1]}, the response from the server was:"
-          puts response.body
+        rescue
+          use_packaged_generators if i == 0
         end
 
       # If a reference to a git repo
       elsif gen.to_s =~ /\.git$/
-        Origen::RevisionControl.new(remote: gen, local: i.to_s).checkout(version: 'master', force: true)
+        begin
+          Origen::RevisionControl.new(remote: gen, local: i.to_s).checkout(version: 'master', force: true)
+        rescue
+          # Better for the user not to crash here, not much we can do otherwise
+        end
 
       # Assume a reference to a folder
       else
-        if File.exist?(gen)
-          FileUtils.cp_r(gen, i.to_s)
-        else
-          puts "Failed to find generator at #{gen}"
+        begin
+          FileUtils.cp_r(gen, i.to_s) if File.exist?(gen)
+        rescue
+          # Better for the user not to crash here, not much we can do otherwise
         end
-
       end
     end
